refactor(core): use new `StatusButton` component
